pyproject.toml
[tool.poetry]
name = "cw7-atomic-habits"
version = "0.1.0"
description = ""
authors = ["TheShadowfish <shadowfish@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
psycopg2 = "^2.9.9"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
ipython = "^8.26.0"
coverage = "^7.6.0"
drf-yasg = "^1.21.7"
stripe = "^10.5.0"
redis = "^5.0.7"
celery = "^5.4.0"
django-celery-beat = "^2.6.0"
django-cors-headers = "^4.4.0"
gunicorn = "^23.0.0"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.1.0"
flake8-docstrings = "^1.7.0"
flake8-builtins = "^2.5.0"
flake8-quotes = "^3.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
